refactor: 상품 상세 화면 컴포넌트 파일 분리
